requirements.txt
django==2.2
tensorflow==2.3.0
keras==2.4.3
Keras-Preprocessing==1.0.5
pillow